+++ example/offline-signing/constants.js
