+++ example/offline-signing/registry.js
