+++ package.json
@@ -34,7 +34,8 @@
     "jsdoc": "^3.6.3",
     "rollup": "2.28.1",
     "typescript": "^4.2.3",
-    "web3": "^1.3.4"
+    "web3": "^1.3.4",
+    "@substrate/txwrapper": "^3.9.2"
   },
   "scripts": {
     "dock-did-example": "npx babel-node example/dock-did.js",
@@ -94,7 +95,6 @@
     "@polkadot/types-known": "4.3.1",
     "@polkadot/util": "6.0.5",
     "@polkadot/util-crypto": "6.0.5",
-    "@substrate/txwrapper": "^3.9.2",
     "axios": "0.21.1",
     "babel-polyfill": "^6.26.0",
     "blake2b": "^2.1.3",
